Prevent models created by guests from saving to DB
When guests previously created an idea or a session, it was saved to the
database. The guest ideas and sessions were deleted whenever a new guest
sign-on was detected.

Changed idea and session add methods to only add models in the view when
logged in as a guest.

Also deleted some commented out code that was not longer being used.