Update guest sessions & user session highlight
Ordering of guest sessions was reversed because session collections sort
by create date. Sessions for guests was previously created through
the server and added to the database. A previous update changed the
session creation to only occur on the client. This resulted in guest
session objects without dates. Didn't add the create date on the
client because the JavaScript date doesn't format easily to the
Postres time stamp. Addressed this by changing the session collection
to sort by id when user is guest.

This change ensured all newly created guest sessions were in order, but
the initial default session was pulling from the database and typically
had a higher id resulting in an incorrect sort. Made a change so the
default session is now created on the client.

The active session wasn't highlighted when a new session was
created for logged in users. This was occuring because
this.updateSessionView was getting called inside a function definition
causing an incorrect reference to window instead of the SessionListView
object.